refact: breakdown notifications lib
